`{insert_records_into => graft_records_onto}_source_chain`
